+++ src/commands/sync.js
@@ -1,0 +1,5 @@
+'use strict';
+
+module.exports = function generate(/* modelName */) {
+  console.log('Not implemented yet.');
+};