+++ src/helpers/ember.js
@@ -1,0 +1,6 @@
+'use strict';
+
+var which = require('npm-which')(process.cwd());
+
+
+module.exports = which.sync('ember');