+++ src/helpers/sails.js
@@ -1,0 +1,5 @@
+'use strict';
+
+var sails = (process.platform === 'win32' ? 'sails.cmd' : 'sails');
+
+module.exports = sails;