+++ tests/runner.js
@@ -4,27 +4,22 @@
 var glob  = require('glob');
 var Mocha = require('mocha');
 var chalk = require('chalk');
-var root;
 
-require('traceur').require.makeDefault(function (filename) {
-// don't transpile our dependencies, just our app
-// The first check is if you develop locally, the second for the globally installed moduel
-  return (filename.indexOf('node_modules') === -1) ||
-    (filename.indexOf('/node_modules/sane-cli/') > -1 && filename.indexOf('/node_modules/sane-cli/node_modules') === -1);
-}, {asyncFunctions: true});
+require('babel/register');
 
 var mocha = new Mocha({
   // For some reason, tests take a long time on Windows (or at least AppVeyor)
-  timeout : (process.platform === 'win32') ? 90000 : 18000,
+  timeout : (process.platform === 'win32') ? 150000 : 40000,
   reporter: 'spec'
 });
 
 // Determine which tests to run based on argument passed to runner
 var arg = process.argv[2];
+var root;
 if (!arg) {
-  var root = 'tests/{unit,acceptance}';
+  root = 'tests/{unit,acceptance}';
 } else if (arg === 'unit') {
-  var root = 'tests/unit';
+  root = 'tests/unit';
 } else {
   root = 'tests/{unit,acceptance}';
 }
@@ -34,6 +29,7 @@ function addFiles(mocha, files) {
 }
 
 addFiles(mocha, '/**/*Test.js');
+// addFiles(mocha, '/**/getAddonTest.js');
 
 mocha.run(function (failures) {
   process.on('exit', function () {