+++ tests/unit/helpers/emberTest.js
@@ -1,4 +1,3 @@
-/*eslint-env node, mocha, es6 */
 'use strict';
 
 var ember = require('../../../lib/helpers/ember');
@@ -7,7 +6,7 @@ var expect = require('chai').expect;
 describe('Unit: ember helper', function () {
 
   it('returns a value', function () {
-    expect(ember).to.be.ok();
+    expect(ember).to.be.ok; // eslint-disable-line no-unused-expressions
   });
 
   it('returns local ember by default', function () {