+++ tests/unit/jshintTest.js
@@ -1,8 +1,0 @@
-// 'use strict';
-
-// require('mocha-jshint')([
-// // Both use async/await, currently unsupported by jshint or eslint
-// // 'tests',
-// // 'lib',
-// 'bin',
-// ]);