+++ tests/unit/tasks/createEmberConfigTest.js
@@ -1,0 +1,14 @@
+'use strict';
+
+var { expect }        = require('chai');
+// var createEmberConfig = require('../../../lib/tasks/generate/createEmberConfig');
+
+describe('createEmberConfig tests', async function() {
+
+  it('creates the correct config', async function() {
+    // NOTE (markus): Create a fixture ember-cli config file, then give new configuration
+    // to add and check that the path as well as the config returned are correct!
+    expect('1').to.equal('1');
+  });
+
+});